requirements.txt
# 糖尿病智能诊断助手依赖
# Flask Web框架
flask==3.0.0
flask-cors==4.0.0

# OpenAI SDK（用于调用百炼API）
openai>=1.3.0

# LlamaIndex RAG框架
llama-index==0.10.30
llama-index-embeddings-dashscope==0.2.0
llama-index-llms-dashscope==0.2.0
llama-index-llms-openai-like==0.1.3

# PDF处理
pypdf==4.0.0
pdfplumber==0.10.3

# Excel处理
pandas==2.1.4
openpyxl==3.1.2

# MySQL连接
pymysql==1.1.0
sqlalchemy==2.0.23

# 向量存储
faiss-cpu==1.7.4

# 数据可视化
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# 定时任务
apscheduler==3.10.4

# 日志
loguru==0.7.2

# 其他工具
python-dotenv==1.0.0
jieba==0.42.1
numpy==1.26.2

